// src/api/generated/model/carResponseStatus.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * MiniCRM API
 * API Documentation for MiniCRM Application
 * OpenAPI spec version: 1.0.0
 */

export type CarResponseStatus = typeof CarResponseStatus[keyof typeof CarResponseStatus];


export const CarResponseStatus = {
  ACTIVE: 'ACTIVE',
  DELETED: 'DELETED',
} as const;